+++ back/src/main/java/com/rouby/user/application/dto/command/ResetPasswordByTokenCommand.java
@@ -1,0 +1,12 @@
+package com.rouby.user.application.dto.command;
+
+import lombok.Builder;
+
+@Builder
+public record ResetPasswordByTokenCommand(
+    String newPassword,
+    String token,
+    String email
+) {
+
+}
