+++ back/src/main/java/com/rouby/user/application/dto/command/ResetPasswordCommand.java
@@ -4,9 +4,8 @@ import lombok.Builder;
 
 @Builder
 public record ResetPasswordCommand(
-    String newPassword,
-    String token,
-    String email
+    String currentPassword,
+    String newPassword
 ) {
 
 }
