+++ back/src/main/java/com/rouby/user/application/service/UserWriteService.java
@@ -5,11 +5,14 @@ import static com.rouby.user.application.exception.UserErrorCode.INVALID_EMAIL_V
 import static com.rouby.user.application.exception.UserErrorCode.PASSWORD_TOKEN_EXPIRED;
 import static com.rouby.user.application.exception.UserErrorCode.USER_NOT_FOUND;
 
+import com.rouby.common.exception.CustomException;
 import com.rouby.user.application.dto.SaveVerificationCodeCommand;
 import com.rouby.user.application.dto.VerifyEmailCommand;
 import com.rouby.user.application.dto.command.CreateUserCommand;
 import com.rouby.user.application.dto.command.FindPasswordCommand;
+import com.rouby.user.application.dto.command.ResetPasswordByTokenCommand;
 import com.rouby.user.application.dto.command.ResetPasswordCommand;
+import com.rouby.user.application.exception.UserErrorCode;
 import com.rouby.user.application.exception.UserException;
 import com.rouby.user.application.service.verification.VerificationEmailCode;
 import com.rouby.user.application.service.verification.VerificationEmailCodeStorage;
@@ -34,7 +37,7 @@ public class UserWriteService {
   @Transactional
   public void create(CreateUserCommand command) {
     ensureEmailNotTaken(command.email());
-    ensureEmailVerified(command.email());
+    //ensureEmailVerified(command.email());
 
     User user = User.create(command.email(), command.password(), passwordEncoder);
     userRepository.save(user);
@@ -82,7 +85,19 @@ public class UserWriteService {
   }
 
   @Transactional
-  public void resetPasswordByToken(ResetPasswordCommand command) {
+  public void resetPassword(Long userId, ResetPasswordCommand command) {
+    User user = userRepository.findById(userId).orElseThrow(() ->
+        new CustomException(UserErrorCode.USER_NOT_FOUND));
+
+    if (!passwordEncoder.matches(command.currentPassword(), user.getPassword())) {
+      throw new CustomException(UserErrorCode.INVALID_PASSWORD);
+    }
+
+    user.updatePassword(passwordEncoder, command.newPassword());
+  }
+
+  @Transactional
+  public void resetPasswordByToken(ResetPasswordByTokenCommand command) {
     User user = userRepository.findByEmail(command.email())
         .orElseThrow(() -> UserException.from(USER_NOT_FOUND));
